Página de confirmação de agendamento

+++ pages/agendado.vue
@@ -1,15 +1,26 @@
 <template>
+  <header class="flex flex-col p-5 w-full h-full" id="cabecalho">
+    <nuxt-link to="/">
+      <div class="p-2">
+        <h1 class="text-4xl md:text-6xl text-center p-2 font-IBM font-bold">
+          TELEMENSAGEM <br />
+          ROSA DE SAROM
+        </h1>
+        <p class="text-center text-lg font-playWrite">
+          💌 Palavras que viram presentes 🎁
+        </p>
+      </div>
+    </nuxt-link>
+  </header>
+  <hr />
   <div class="bg-red-600 opacity-70 p-5 text-2xl rounded-xl text-center">
     <div class="flex flex-col items-center p-5 text-center">
       <h1 class="font-bold text-5xl">Agendamento realizado!</h1>
-      <p class="text-2xl font-bold underline" v-if="propModelo === 'Ao Vivo'">
+      <p class="text-2xl font-bold underline">
         O pagamento é feito no local da mensagem ao vivo
       </p>
       <p class="text-2xl">Agora basta confiar e aguardar</p>
       <p class="text-2xl">Faltam {{ timeRemaining }} até a data da mensagem</p>
-      <p class="text-2xl">
-        Quando esta tela sumir, quer dizer que sua mensagem foi enviada
-      </p>
       <div>
         <img src="/images/check-icon-gif.gif" alt="Check" class="w-40" />
       </div>
@@ -59,23 +70,15 @@
 </template>
 
 <script lang="ts" setup>
-const emit = defineEmits(["agendamentoAtivo"]);
+import { useFormStore } from "~/store/userFormStore";
 
-const props = defineProps<{
-  propModelo: string;
-  propSentTime: string;
-  propSentDate: string;
-}>();
-
-function updateTimeRemaining(): string {
+const timeRemaining = ref<string>("");
+function updateTimeRemaining(): void {
   const now = new Date();
-  const target = new Date(props.propSentDate);
+  const target = new Date(useFormStore().formData.hora);
   const diff: number = Math.abs(target.getTime() - now.getTime());
 
-  if (diff <= 0) {
-    localStorage.clear();
-    emit("agendamentoAtivo", false);
-  } else emit("agendamentoAtivo", true);
+  if (diff <= 0) localStorage.clear();
 
   const days: number = Math.floor(diff / (1000 * 60 * 60 * 24));
   const hours: number = Math.floor(
@@ -84,20 +87,19 @@ function updateTimeRemaining(): string {
   const minutes: number = Math.floor((diff % (1000 * 60 * 60)) / (1000 * 60));
   const seconds: number = Math.floor((diff % (1000 * 60)) / 1000);
 
-  return `${days}d ${hours}h ${minutes}m ${seconds}s`;
+  timeRemaining.value = `${days}d ${hours}h ${minutes}m ${seconds}s`;
 }
 
-// Atualiza a cada segundo
-const timeRemaining = ref<string>("");
-onMounted(() => {
-  timeRemaining.value = updateTimeRemaining(); // Atualiza imediatamente ao montar
-  intervalId = setInterval(
-    () => (timeRemaining.value = updateTimeRemaining()),
-    1000,
-  );
+let intervalId: any;
+// Verifica se já houve agendamento ou não
+onBeforeMount(() => {
+  const isAgendadoActive = localStorage.getItem("agendado");
+  if (isAgendadoActive === null) return useRouter().replace("/agendamento");
+
+  updateTimeRemaining(); // Atualiza imediatamente ao montar
+  intervalId = setInterval(updateTimeRemaining, 1000);
 });
 
-let intervalId: any;
 onUnmounted(() => {
   if (intervalId !== undefined) clearInterval(intervalId); // Limpa o intervalo quando o componente é desmontado
 });
